HTBHF-413 refactor are you pregnant unit tests
