+++ src/web/routes/application/are-you-pregnant/test/validate.test.js
@@ -1,71 +1,10 @@
 const test = require('tape')
-const { applyExpressValidation } = require('../common/test/apply-express-validation')
-const { validate, addExpectedDeliveryDateToBody, validateExpectedDeliveryDate } = require('./validate')
-const { dateAsString } = require('../common/formatters')
-const { YES, NO } = require('./constants')
+const { addExpectedDeliveryDateToBody, validateExpectedDeliveryDate } = require('../validate')
+const { dateAsString } = require('../../common/formatters')
+const { YES, NO } = require('../constants')
 
 const res = {}
 const next = () => {}
-
-test('validation middleware passes with valid body', async (t) => {
-  const sixMonthsInFuture = dateAsString({ monthAdjustment: 6 }).split('-')
-
-  const req = {
-    t: string => string,
-    body: {
-      areYouPregnant: YES,
-      'expectedDeliveryDate-day': sixMonthsInFuture[2],
-      'expectedDeliveryDate-month': sixMonthsInFuture[1],
-      'expectedDeliveryDate-year': sixMonthsInFuture[0]
-    }
-  }
-
-  const result = await applyExpressValidation(req, validate)
-  t.equal(result.isEmpty(), true)
-  t.end()
-})
-
-test('validation middleware errors for are you pregnant field', async (t) => {
-  const sixMonthsInFuture = dateAsString({ monthAdjustment: 6 }).split('-')
-
-  const req = {
-    t: string => string,
-    body: {
-      areYouPregnant: 'maybe',
-      'expectedDeliveryDate-day': sixMonthsInFuture[2],
-      'expectedDeliveryDate-month': sixMonthsInFuture[1],
-      'expectedDeliveryDate-year': sixMonthsInFuture[0]
-    }
-  }
-
-  const result = await applyExpressValidation(req, validate)
-  const error = result.array()[0]
-  t.equal(result.isEmpty(), false, 'should have errors')
-  t.equal(error.param, 'areYouPregnant', 'error should be associated with correct field')
-  t.equal(error.msg, 'validation:selectYesOrNo', 'error should have correct message')
-  t.end()
-})
-
-test('validation middleware errors for are you pregnant field', async (t) => {
-  const twelveMonthsInFuture = dateAsString({ monthAdjustment: 12 }).split('-')
-
-  const req = {
-    t: string => string,
-    body: {
-      areYouPregnant: YES,
-      'expectedDeliveryDate-day': twelveMonthsInFuture[2],
-      'expectedDeliveryDate-month': twelveMonthsInFuture[1],
-      'expectedDeliveryDate-year': twelveMonthsInFuture[0]
-    }
-  }
-
-  const result = await applyExpressValidation(req, validate)
-  const error = result.array()[0]
-  t.equal(result.isEmpty(), false, 'should have errors')
-  t.equal(error.param, 'expectedDeliveryDate', 'error should be associated with correct field')
-  t.equal(error.msg, 'validation:expectedDeliveryDateInvalidTooFarInFuture', 'error should have correct message')
-  t.end()
-})
 
 test('validateExpectedDeliveryDate()', (t) => {
   const req = {
